refactor: 폴더명 수정으로 인한 파일 경로 관련 리팩토링
