+++ src/components/tablings/InputName.tsx
@@ -20,7 +20,7 @@ const InputName: React.FC<InputNameProps> = ({ value, onChange }) => {
     <>
       <div className="text-xs">이름</div>
       <div className="h-11 w-full flex flex-row items-center py-2.5 gap-2.5">
-        <img src="/icons/tabling/person.svg" className="w-6 h-6" />
+        <img src="/icons/tablings/person.svg" className="w-6 h-6" />
         <input
           className="flex-1 focus:outline-none bg-inherit"
           value={value}
