+++ src/components/tablings/InputPhoneNum.tsx
@@ -25,7 +25,7 @@ const InputPhoneNum: React.FC<InputPhoneNumProps> = ({ value, onChange }) => {
     <>
       <div className="text-xs">전화번호</div>
       <div className="h-11 w-full flex flex-row items-center py-2.5 gap-2.5">
-        <img src="/icons/tabling/phone.svg" className="w-6 h-6" />
+        <img src="/icons/tablings/phone.svg" className="w-6 h-6" />
         <input
           className="flex-1 focus:outline-none bg-inherit"
           type="tel"
