+++ src/components/tablings/NoBooth.tsx
